Find some bugs and go on debugging

+++ src/include/storage/page/b_plus_tree_internal_page.h
@@ -44,7 +44,7 @@ class BPlusTreeInternalPage : public BPlusTreePage {
   void SetValueAt(int index, const page_id_t &value);
   auto ValueAt(int index) const -> ValueType;
   auto Bisect(KeyType const &key, page_id_t *page_id, KeyComparator const &comparator) const -> bool;
-  auto BisectPosition(KeyType const &key, KeyComparator const &comparator) const -> size_t;
+  auto BisectPosition(KeyType const &key, KeyComparator const &comparator) const -> int;
   auto InsertAt(int index, KeyType const &key, page_id_t const &page_id) -> void;
   auto IncrementSize() -> void;
   auto DecrementSize() -> void;
